+++ src/connect4/search/compressed_encoding.c
@@ -7,17 +7,17 @@
 void initialise_variables(nodeindex_t (**X)[2], uint32_t width, uint32_t height) {
     // Second, cells: order column, row, player, board
     if (!ALLOW_ROW_ORDER || width >= height) {
-        for (int col = 0; col < width; col++) {
-            for (int row = 0; row < height + 1; row++) {
-                for (int board = 0; board < 2; board++) {
+        for (uint8_t col = 0; col < width; col++) {
+            for (uint8_t row = 0; row < height + 1; row++) {
+                for (uint8_t board = 0; board < 2; board++) {
                     X[col][row][board] = create_variable();
                 }
             }
         }
     } else {
-        for (int row = 0; row < height + 1; row++) {
-            for (int col = 0; col < width; col++) {
-                for (int board = 0; board < 2; board++) {
+        for (uint8_t row = 0; row < height + 1; row++) {
+            for (uint8_t col = 0; col < width; col++) {
+                for (uint8_t board = 0; board < 2; board++) {
                     X[col][row][board] = create_variable();
                 }
             }
@@ -26,8 +26,8 @@ void initialise_variables(nodeindex_t (**X)[2], uint32_t width, uint32_t height)
 }
 
 void initialise_variable_sets(nodeindex_t (**X)[2], nodeindex_t stm0, nodeindex_t stm1, uint32_t width, uint32_t height, variable_set_t* vars_board0, variable_set_t* vars_board1) {
-    for (int col = 0; col < width; col++) {
-        for (int row = 0; row < height + 1; row++) {
+    for (uint8_t col = 0; col < width; col++) {
+        for (uint8_t row = 0; row < height + 1; row++) {
             add_variable(vars_board0, X[col][row][0]);
             add_variable(vars_board1, X[col][row][1]);
         }
@@ -37,9 +37,9 @@ void initialise_variable_sets(nodeindex_t (**X)[2], nodeindex_t stm0, nodeindex_
 }
 
 
-nodeindex_t trans_move(int col, int row, int player, int current_board, nodeindex_t stm0, nodeindex_t stm1, nodeindex_t (**X)[2], uint32_t width, uint32_t height) {
-    // int opponent = player == 0 ? 1 : 0;
-    int next_board = current_board == 0 ? 1 : 0;
+nodeindex_t trans_move(uint8_t col, uint8_t row, uint8_t player, uint8_t current_board, nodeindex_t stm0, nodeindex_t stm1, nodeindex_t (**X)[2], uint32_t width, uint32_t height) {
+    // uint8_t opponent = player == 0 ? 1 : 0;
+    uint8_t next_board = current_board == 0 ? 1 : 0;
     
     // precondition (depends variables on of current board)
     nodeindex_t pre;
@@ -52,7 +52,7 @@ nodeindex_t trans_move(int col, int row, int player, int current_board, nodeinde
     // cell is high
     pre = and(pre, X[col][row][current_board]);
     // everything above cell is low
-    for (int r = row+1; r < height + 1; r++) {
+    for (uint8_t r = row+1; r < height + 1; r++) {
         pre = and(pre, not(X[col][r][current_board]));
     }
 
@@ -75,8 +75,8 @@ nodeindex_t trans_move(int col, int row, int player, int current_board, nodeinde
 
     // frame: every other cell stays the same
     nodeindex_t frame = ONEINDEX;
-    for (int c = 0; c < width; c++) {
-        for (int r = 0; r < height + 1; r++) {
+    for (uint8_t c = 0; c < width; c++) {
+        for (uint8_t r = 0; r < height + 1; r++) {
             if ((c == col) && (r == row)) {
                 continue;
             }
@@ -92,11 +92,11 @@ nodeindex_t trans_move(int col, int row, int player, int current_board, nodeinde
 }
 
 
-nodeindex_t get_trans(nodeindex_t (**X)[2], nodeindex_t stm0, nodeindex_t stm1, uint32_t width, uint32_t height, int board) {
+nodeindex_t get_trans(nodeindex_t (**X)[2], nodeindex_t stm0, nodeindex_t stm1, uint32_t width, uint32_t height, uint8_t board) {
     nodeindex_t trans = ZEROINDEX;
-    for (int col = 0; col < width; col++) {
-        for (int row = 0; row < height; row++) { // it is enough to go to row < height
-            for (int player = 0; player < 2; player++) {
+    for (uint8_t col = 0; col < width; col++) {
+        for (uint8_t row = 0; row < height; row++) { // it is enough to go to row < height
+            for (uint8_t player = 0; player < 2; player++) {
                 trans = or(trans, trans_move(col, row, player, board, stm0, stm1, X, width, height));
             }
         }
@@ -107,18 +107,18 @@ nodeindex_t get_trans(nodeindex_t (**X)[2], nodeindex_t stm0, nodeindex_t stm1, 
 nodeindex_t connect4_start(nodeindex_t stm0, nodeindex_t (**X)[2], uint32_t width, uint32_t height) {
     nodeindex_t s = ONEINDEX;
     s = and(s, stm0);
-    for (int c = 0; c < width; c++) {
+    for (uint8_t c = 0; c < width; c++) {
         s = and(s, X[c][0][0]); // bottom row is high
-        for (int r = 1; r < height + 1; r++) {
+        for (uint8_t r = 1; r < height + 1; r++) {
             s = and(s, not(X[c][r][0])); // everthing else is slow
         }
     }
     return s;
 }
 
-nodeindex_t is_valid_cell(nodeindex_t (**X)[2], uint32_t height, int col, int row, int board) {
+nodeindex_t is_valid_cell(nodeindex_t (**X)[2], uint32_t height, uint8_t col, uint8_t row, uint8_t board) {
     nodeindex_t a = ZEROINDEX;
-    for (int r = row+1; r < height + 1; r++) {
+    for (uint8_t r = row+1; r < height + 1; r++) {
         a = or(a, X[col][r][board]);
     }
     return a;
@@ -126,7 +126,7 @@ nodeindex_t is_valid_cell(nodeindex_t (**X)[2], uint32_t height, int col, int ro
 
 // Subtracts all positions from current which are terminal, i.e. four in a row, column or diagonal
 // Subtraction is performed iteratively and also performs GC.
-nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, int player, nodeindex_t (**X)[2], uint32_t width, uint32_t height, int gc_level, bool subtract) {
+nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, uint8_t board, uint8_t player, nodeindex_t (**X)[2], uint32_t width, uint32_t height, int gc_level, bool subtract) {
     nodeindex_t a;
     nodeindex_t x;
     nodeindex_t intersection = ZEROINDEX;
@@ -135,12 +135,12 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
 
     // COLUMN
     if (height >= 4) {
-        for (int col = 0; col < width; col++) {
-            for (int row = 0; row <= height - 4; row++) {
+        for (uint8_t col = 0; col < width; col++) {
+            for (uint8_t row = 0; row <= height - 4; row++) {
                 // encodes that there are four stones in a column for player starting at (col,row)
                 a = ONEINDEX;
                 keepalive_ix(a);
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     x = (player == 0) ? X[col][row + i][board] : not(X[col][row + i][board]);
                     x = and(x, is_valid_cell(X, height, col, row + i, board));
                     reassign_and_keepalive(&a, and(a, x));
@@ -172,12 +172,12 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
 
     // ROW
     if (width >= 4) {
-        for (int row = 0; row < height; row++) {
-            for (int col = 0; col <= width - 4; col++) {
+        for (uint8_t row = 0; row < height; row++) {
+            for (uint8_t col = 0; col <= width - 4; col++) {
                 // encodes that there are four stones in a row for player starting at (col,row)
                 a = ONEINDEX;
                 keepalive_ix(a);
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     x = (player == 0) ? X[col + i][row][board] : not(X[col + i][row][board]);
                     x = and(x, is_valid_cell(X, height, col + i, row, board));
                     reassign_and_keepalive(&a, and(a, x));
@@ -202,12 +202,12 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
 
     if (height >= 4 && width >= 4) {
         // DIAG ascending
-        for (int col = 0; col <= width - 4; col++) {
-            for (int row = 0; row <= height - 4; row++) {
+        for (uint8_t col = 0; col <= width - 4; col++) {
+            for (uint8_t row = 0; row <= height - 4; row++) {
                 // encodes that there are four stones in a ascending diagonal for player starting at (col,row)
                 a = ONEINDEX;
                 keepalive_ix(a);
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     x = (player == 0) ? X[col + i][row + i][board] : not(X[col + i][row + i][board]);
                     x = and(x, is_valid_cell(X, height, col + i, row + i, board));
                     reassign_and_keepalive(&a, and(a, x));
@@ -230,12 +230,12 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
         }
 
         // DIAG descending
-        for (int col = 3; col < width; col++) {
-            for (int row = 0; row <= height - 4; row++) {
+        for (uint8_t col = 3; col < width; col++) {
+            for (uint8_t row = 0; row <= height - 4; row++) {
                 // encodes that there are four stones in a descending diagonal for player starting at (col+3,row)
                 a = ONEINDEX;
                 keepalive_ix(a);
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     x = (player == 0) ? X[col - i][row + i][board] : not((X[col - i][row + i][board]));
                     x = and(x, is_valid_cell(X, height, col - i, row + i, board));
                     reassign_and_keepalive(&a, and(a, x));
@@ -267,9 +267,9 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
     }
 }
 
-inline nodeindex_t connect4_subtract_term(nodeindex_t current, int board, int player, nodeindex_t (**X)[2], uint32_t width, uint32_t height, int gc_level) {
+inline nodeindex_t connect4_subtract_term(nodeindex_t current, uint8_t board, uint8_t player, nodeindex_t (**X)[2], uint32_t width, uint32_t height, int gc_level) {
     return connect4_subtract_or_intersect_term(current, board, player, X, width, height, gc_level, true);
 }
-inline nodeindex_t connect4_intersect_term(nodeindex_t current, int board, int player, nodeindex_t (**X)[2], uint32_t width, uint32_t height, int gc_level) {
+inline nodeindex_t connect4_intersect_term(nodeindex_t current, uint8_t board, uint8_t player, nodeindex_t (**X)[2], uint32_t width, uint32_t height, int gc_level) {
     return connect4_subtract_or_intersect_term(current, board, player, X, width, height, gc_level, false);
 }
